# Conflicts:
#	ClassesGUI.py

+++ ClassesGUI.py
@@ -3,6 +3,7 @@ from tkinter import *
 from tkinter import messagebox
 import time
 import sqlite3
+from datetime import datetime
 
 from ManagerFunctions import MenuEdit, ManagerFunctions
 
@@ -87,7 +88,8 @@ class CookWindow():
         clock_frame.pack()
 
         # clock in button
-        self.ClockIn_Button = Button(clock_frame, text="Clock In", font=('Segoe UI Light', 16), width=10)
+        self.ClockIn_Button = Button(clock_frame, text="Clock In", font=('Segoe UI Light', 16), width=10,
+                                     command=self.clockIn)
         self.ClockIn_Button.pack(side=LEFT, padx=5, pady=10)
 
         # clock out button
@@ -100,7 +102,8 @@ class CookWindow():
         check_order_frame.pack()
 
         # get check button
-        self.GetCheck_Button = Button(check_order_frame, text="Get Check", font=('Segoe UI Light', 16), width=10)
+        self.GetCheck_Button = Button(check_order_frame, text="Get Check", font=('Segoe UI Light', 16), width=10,
+                                      command=self.getCheck)
         self.GetCheck_Button.pack(side=LEFT, padx=5, pady=10)
 
         # track order time button
@@ -124,12 +127,16 @@ class CookWindow():
         destroyDisplay()
         mywindow = MainWindow(root)
 
+    def CookBack_Pushed(self):
+        destroyDisplay()
+        cookwindow = CookWindow()
+
     def clockOut(self):
         destroyDisplay()
-        frame = Frame()
+        frame = Frame(bg="#2d2d2d")
         frame.pack()
-        root.geometry("150x100")
-        header = Label(frame, text="Enter Pin to Clock Out")
+        root.geometry("250x100+700+300")
+        header = Label(frame, text="Enter Pin to Clock Out", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
         header.grid(row=1, column=1, ipadx="10")
         pin_field = Entry(frame)
         pin_field.grid(row=2, column=1, ipadx="10")
@@ -138,13 +145,100 @@ class CookWindow():
         def submit():
             pin = pin_field.get()
             if (pin != ""):
-                clockOutFile = open(pin + ".txt", "a+")
-                clockOutFile.write("\nOut " + timestr)
-                clockOutFile.close
-                destroyDisplay()
-                mywindow = MainWindow(root)
+                getClockState = open(pin + ".txt", "r")
+                isClockedIn = getClockState.read()
+                if isClockedIn == "IN":
+                    clockOutFile = open(pin + ".txt", "w")
+                    clockOutFile.write(timestr)
+                    clockOutFile.close
+                    clockOutFile = open(pin + ".txt", "r+") # go back and open file for IN/OUT flag
+                    clockOutFile.seek(0) # go to line 1 of file
+                    clockOutFile.write("OUT\n") # signal clock OUT
+                    clockOutFile.close
+                    destroyDisplay()
+                    mywindow = MainWindow(root)
+                else:
+                    print ("Staff must be clocked-in to clock-out")
 
-        self.SubmitButton = Button(frame, text="Clock Out", command=submit)
+        self.SubmitButton = Button(frame, text="Clock Out", font=('Segoe UI Light', 12), command=submit)
+        self.SubmitButton.grid(row=3, column=1, ipadx="10")
+
+    def clockIn(self):
+        destroyDisplay()
+        frame = Frame(bg="#2d2d2d")
+        frame.pack()
+        root.geometry("250x100+700+300")
+        header = Label(frame, text="Enter Pin to Clock In", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
+        header.grid(row=1, column=1, ipadx="10")
+        pin_field = Entry(frame)
+        pin_field.grid(row=2, column=1, ipadx="10")
+        timestr = time.strftime("%Y %m %d- %H %M %S")
+
+        def submit():
+            pin = pin_field.get()
+            if (pin != ""):
+                getClockState = open(pin + ".txt", "r")
+                isClockedOut = getClockState.read()
+                if isClockedOut == "OUT":
+                    clockOutFile = open(pin + ".txt", "w")
+                    clockOutFile.write(timestr)
+                    clockOutFile.close
+                    clockOutFile = open(pin + ".txt", "r+") # go back and open file for IN/OUT flag
+                    clockOutFile.seek(0) # go to line 1 of file
+                    clockOutFile.write("IN\n") # signal clock IN
+                    clockOutFile.close
+                    destroyDisplay()
+                    cookWindow = CookWindow()
+                else:
+                    print ("Cook must be clocked-out to clock-in")
+
+        self.SubmitButton = Button(frame, text="Clock In", font=('Segoe UI Light', 12), command=submit)
+        self.SubmitButton.grid(row=3, column=1, ipadx="10")
+
+    def getCheck(self):
+        destroyDisplay()
+        frame = Frame(bg="#2d2d2d")
+        frame.pack()
+        root.geometry("250x140+700+300")
+        header = Label(frame, text="Enter Pin to get check", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
+        header.grid(row=1, column=1, ipadx="10")
+        pin_field = Entry(frame)
+        pin_field.grid(row=2, column=1, ipadx="10")
+        self.BackButton = Button(frame, text="Back", font=('Segoe UI Light', 12), command=self.CookBack_Pushed)
+        self.BackButton.grid(row=4, column=1, ipadx="10")
+
+        def submit():
+            pin = pin_field.get()
+            if (pin != ""):
+                clockOutIn = open(pin + ".txt", "r")
+                clockOutOut = open(pin + ".txt", "r")
+
+                InTime = clockOutIn.read()
+                OutTime = clockOutOut.read()
+                print(InTime)
+                print(OutTime)
+
+                tdelta = datetime.strptime(OutTime, "%Y %m %d- %H %M %S") - datetime.strptime(InTime,
+                                                                                              "%Y %m %d- %H %M %S")
+                TotalMoney = round(tdelta.total_seconds() * .002, 2)
+                print(tdelta.total_seconds())
+                print(TotalMoney)
+
+                clockOutIn.close
+                clockOutOut.close
+
+                destroyDisplay()
+                frame = Frame(bg="#2d2d2d")
+                frame.pack()
+                root.geometry("260x100+700+300")
+                header = Label(frame, text="Check for $" + str(TotalMoney) + " received", font=('Segoe UI Light', 16),
+                               bg="#2d2d2d", fg="#ffffff")
+                header.grid(row=1, column=1, ipadx="10")
+
+                self.BackButton = Button(frame, text="Back", width=5, font=('Segoe UI Light', 12), command=self.CookBack_Pushed)
+                self.BackButton.grid(row=3, column=1, ipady="5")
+
+        self.SubmitButton = Button(frame, text="Get Check", font=('Segoe UI Light', 12), command=submit)
         self.SubmitButton.grid(row=3, column=1, ipadx="10")
 
 
@@ -187,7 +281,8 @@ class StaffWindow():
         clock_in_out.pack()
 
         # clock in button
-        self.ClockIn_Button = Button(clock_in_out, text="Clock In", font=('Segoe UI Light', 16), width=10)
+        self.ClockIn_Button = Button(clock_in_out, text="Clock In", font=('Segoe UI Light', 16), width=10,
+                                     command=self.clockIn)
         self.ClockIn_Button.pack(side=LEFT, padx=5, pady=10)
 
         # clock out button
@@ -214,10 +309,10 @@ class StaffWindow():
 
     def clockOut(self):
         destroyDisplay()
-        frame = Frame()
+        frame = Frame(bg="#2d2d2d")
         frame.pack()
-        root.geometry("150x100")
-        header = Label(frame, text="Enter Pin to Clock Out")
+        root.geometry("250x100+700+300")
+        header = Label(frame, text="Enter Pin to Clock Out", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
         header.grid(row=1, column=1, ipadx="10")
         pin_field = Entry(frame)
         pin_field.grid(row=2, column=1, ipadx="10")
@@ -226,13 +321,54 @@ class StaffWindow():
         def submit():
             pin = pin_field.get()
             if (pin != ""):
-                clockOutFile = open(pin + ".txt", "a+")
-                clockOutFile.write("\nOut " + timestr)
-                clockOutFile.close
-                destroyDisplay()
-                mywindow = MainWindow(root)
+                getClockState = open(pin + ".txt", "r")
+                isClockedIn = getClockState.read()
+                if isClockedIn == "IN":
+                    clockOutFile = open(pin + ".txt", "a+")
+                    clockOutFile.write("\nOut " + timestr)
+                    clockOutFile.close
+                    clockOutFile = open(pin + ".txt", "r+") # go back and open file for IN/OUT flag
+                    clockOutFile.seek(0) # go to line 1 of file
+                    clockOutFile.write("OUT\n") # signal clock OUT
+                    clockOutFile.close
+                    destroyDisplay()
+                    mywindow = MainWindow(root)
+                else:
+                    print ("Staff must be clocked-in to clock-out")
 
-        self.SubmitButton = Button(frame, text="Clock Out", command=submit)
+        self.SubmitButton = Button(frame, text="Clock Out", font=('Segoe UI Light', 12), command=submit)
+        self.SubmitButton.grid(row=3, column=1, ipadx="10")
+
+    def clockIn(self):
+        destroyDisplay()
+        frame = Frame(bg="#2d2d2d")
+        frame.pack()
+        root.geometry("250x100+700+300")
+        header = Label(frame, text="Enter Pin to Clock In", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
+        header.grid(row=1, column=1, ipadx="10")
+        pin_field = Entry(frame)
+        pin_field.grid(row=2, column=1, ipadx="10")
+        timestr = time.strftime("%Y %m %d- %H %M %S")
+
+        def submit():
+            pin = pin_field.get()
+            if (pin != ""):
+                getClockState = open(pin + ".txt", "r")
+                isClockedOut = getClockState.read()
+                if isClockedOut == "OUT":
+                    clockOutFile = open(pin + ".txt", "a+")
+                    clockOutFile.write("\nIn " + timestr)
+                    clockOutFile.close
+                    clockOutFile = open(pin + ".txt", "r+") # go back and open file for IN/OUT flag
+                    clockOutFile.seek(0) # go to line 1 of file
+                    clockOutFile.write("IN\n") # signal clock IN
+                    clockOutFile.close
+                    destroyDisplay()
+                    staffwindow = StaffWindow()
+                else:
+                    print ("Staff must be clocked-out to clock-in")
+
+        self.SubmitButton = Button(frame, text="Clock In", font=('Segoe UI Light', 12), command=submit)
         self.SubmitButton.grid(row=3, column=1, ipadx="10")
 
 
@@ -255,7 +391,6 @@ class ManagerWindow():
         new_close_frame = Frame(root, bg="#2d2d2d")
         new_close_frame.pack()
 
-        # new table button
         self.MenuEdit_Button = Button(new_close_frame, text="Manager Functions", font=('Segoe UI Light', 16), width=15)
         self.MenuEdit_Button.bind("<Button>", lambda e: ManagerFunctions(root, self))
         self.MenuEdit_Button.pack(side=LEFT, padx=5, pady=10)
@@ -280,20 +415,16 @@ class ManagerWindow():
         self.CheckOut_Button = Button(get_out_check, text="Check Out", font=('Segoe UI Light', 16), width=10)
         self.CheckOut_Button.pack(side=LEFT, padx=5, pady=10)
 
-        # clock in button
-        self.ClockIn_Button = Button(get_out_check, text="Clock In", font=('Segoe UI Light', 16), width=10)
-        self.ClockIn_Button.pack(side=LEFT, padx=5, pady=10)
-
         # clock in out frame
         clock_in_out = Frame(root, bg="#2d2d2d")
         clock_in_out.pack()
 
-        # clock out button
-        self.ClockOut_Button = Button(clock_in_out, text="Clock Out", font=('Segoe UI Light', 16), width=10,
-                                      command=self.clockOut)
-        self.ClockOut_Button.pack(side=LEFT, padx=5, pady=10)
+        # clock in button
+        self.ClockIn_Button = Button(clock_in_out, text="Clock In", font=('Segoe UI Light', 16), width=10,
+                                     command=self.clockIn)
+        self.ClockIn_Button.pack(side=LEFT, padx=5, pady=10)
 
-        # manager functions button
+        # clock out button
         self.ClockOut_Button = Button(clock_in_out, text="Clock Out", font=('Segoe UI Light', 16), width=10,
                                       command=self.clockOut)
         self.ClockOut_Button.pack(side=LEFT, padx=5, pady=10)
@@ -313,7 +444,7 @@ class ManagerWindow():
 
     # Put Functions Under This Comment
 
-    def showMenu(self):
+    def showMenu():
         print("Manager Menu")
 
     print("1) clock in")
@@ -353,10 +484,10 @@ class ManagerWindow():
 
     def clockOut(self):
         destroyDisplay()
-        frame = Frame()
+        frame = Frame(bg="#2d2d2d")
         frame.pack()
-        root.geometry("150x100")
-        header = Label(frame, text="Enter Pin to Clock Out")
+        root.geometry("250x100+700+300")
+        header = Label(frame, text="Enter Pin to Clock Out", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
         header.grid(row=1, column=1, ipadx="10")
         pin_field = Entry(frame)
         pin_field.grid(row=2, column=1, ipadx="10")
@@ -365,13 +496,54 @@ class ManagerWindow():
         def submit():
             pin = pin_field.get()
             if (pin != ""):
-                clockOutFile = open(pin + ".txt", "a+")
-                clockOutFile.write("\nOut " + timestr)
-                clockOutFile.close
-                destroyDisplay()
-                mywindow = MainWindow(root)
+                getClockState = open(pin + ".txt", "r")
+                isClockedIn = getClockState.read()
+                if isClockedIn == "IN":
+                    clockOutFile = open(pin + ".txt", "a+")
+                    clockOutFile.write("\nOut " + timestr)
+                    clockOutFile.close
+                    clockOutFile = open(pin + ".txt", "r+") # go back and open file for IN/OUT flag
+                    clockOutFile.seek(0) # go to line 1 of file
+                    clockOutFile.write("OUT\n") # signal clock OUT
+                    clockOutFile.close
+                    destroyDisplay()
+                    mywindow = MainWindow(root)
+                else:
+                    print ("Manager must be clocked-in to clock-out")
 
-        self.SubmitButton = Button(frame, text="Clock Out", command=submit)
+        self.SubmitButton = Button(frame, text="Clock Out", font=('Segoe UI Light', 12), command=submit)
+        self.SubmitButton.grid(row=3, column=1, ipadx="10")
+
+    def clockIn(self):
+        destroyDisplay()
+        frame = Frame(bg="#2d2d2d")
+        frame.pack()
+        root.geometry("250x100+700+300")
+        header = Label(frame, text="Enter Pin to Clock In", font=('Segoe UI Light', 16), bg="#2d2d2d", fg="#ffffff")
+        header.grid(row=1, column=1, ipadx="10")
+        pin_field = Entry(frame)
+        pin_field.grid(row=2, column=1, ipadx="10")
+        timestr = time.strftime("%Y %m %d- %H %M %S")
+
+        def submit():
+            pin = pin_field.get()
+            if (pin != ""):
+                getClockState = open(pin + ".txt", "r")
+                isClockedOut = getClockState.read()
+                if isClockedOut == "OUT":
+                    clockOutFile = open(pin + ".txt", "a+")
+                    clockOutFile.write("\nIn " + timestr)
+                    clockOutFile.close
+                    clockOutFile = open(pin + ".txt", "r+") # go back and open file for IN/OUT flag
+                    clockOutFile.seek(0) # go to line 1 of file
+                    clockOutFile.write("IN\n") # signal clock IN
+                    clockOutFile.close
+                    destroyDisplay()
+                    managerwindow = ManagerWindow()
+                else:
+                    print ("Manager must be clocked-out to clock-in")
+
+        self.SubmitButton = Button(frame, text="Clock In", font=('Segoe UI Light', 12), command=submit)
         self.SubmitButton.grid(row=3, column=1, ipadx="10")
